chore: update React sample app
